Refactor LoginForm into its own component
